# api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PudimFinance API",
    "description": "Personal finance application API. Immutable double-entry ledger, event sourcing, real-time reporting, receipt scanning and price tracking.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/reports/cash-flow": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Dashboard cash-flow totals grouped at the resolution needed by each chart view.",
        "operationId": "cash_flow",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "description": "Start date (inclusive)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "End date (inclusive)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "granularity",
            "in": "query",
            "description": "Aggregation: day, week, or month",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cash-flow totals for the requested periods",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid date range or granularity"
          }
        }
      }
    },
    "/api/reports/category-breakdown": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Category spending breakdown for a date range.",
        "operationId": "category_breakdown",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "description": "Start date (ISO). Default: first day of current month",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "End date (ISO). Default: today",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category spending breakdown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryBreakdownResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid date range"
          }
        }
      }
    },
    "/api/reports/monthly": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Monthly income/expense summary over a date range.",
        "operationId": "monthly_report",
        "parameters": [
          {
            "name": "start_year",
            "in": "query",
            "description": "Start year (default: 6 months ago)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "start_month",
            "in": "query",
            "description": "Start month 1-12",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "end_year",
            "in": "query",
            "description": "End year (default: current)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "end_month",
            "in": "query",
            "description": "End month 1-12",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "account_id",
            "in": "query",
            "description": "Restrict to a single source account (payment method) UUID",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Monthly income/expense summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyReportResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          }
        }
      }
    },
    "/api/reports/trends": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Monthly trends (income, expense, net) over the last N months.",
        "operationId": "trends",
        "parameters": [
          {
            "name": "months",
            "in": "query",
            "description": "Number of months (default 6, max 12)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Monthly trends",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid months parameter"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CashFlowPoint": {
        "type": "object",
        "description": "A single cash-flow period, grouped by day, week, or month.",
        "required": [
          "period_start",
          "income_total",
          "expense_total",
          "balance"
        ],
        "properties": {
          "balance": {
            "type": "string",
            "description": "Balance = income − expense."
          },
          "expense_total": {
            "type": "string",
            "description": "Total expenses for the period."
          },
          "income_total": {
            "type": "string",
            "description": "Total income for the period."
          },
          "period_start": {
            "type": "string",
            "format": "date",
            "description": "First date of the period. Weeks start on Monday; months start on day one."
          }
        }
      },
      "CashFlowResponse": {
        "type": "object",
        "description": "Response for the dashboard cash-flow chart.",
        "required": [
          "points"
        ],
        "properties": {
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CashFlowPoint"
            },
            "description": "Continuous cash-flow points in chronological order."
          }
        }
      },
...
      "CategoryBreakdownItem": {
        "type": "object",
        "description": "A single category's aggregated totals for the category-breakdown report.",
        "required": [
          "total",
          "percentage",
          "transaction_count"
        ],
        "properties": {
          "category_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Category UUID (null when uncategorised)."
          },
          "category_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Category name."
          },
          "color": {
            "type": [
              "string",
              "null"
            ],
            "description": "Hex color."
          },
          "icon": {
            "type": [
              "string",
              "null"
            ],
            "description": "Icon identifier."
          },
          "percentage": {
            "type": "string",
            "description": "Percentage of all expenses for the period."
          },
          "total": {
            "type": "string",
            "description": "Total amount for the category."
          },
          "transaction_count": {
            "type": "integer",
            "format": "int64",
            "description": "Number of transactions in this category."
          }
        }
      },
      "CategoryBreakdownResponse": {
        "type": "object",
        "description": "Response for the category-breakdown report.",
        "required": [
          "categories",
          "start_date",
          "end_date"
        ],
        "properties": {
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryBreakdownItem"
            },
            "description": "Per-category totals, sorted by total descending."
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "description": "Date range used (ISO dates)."
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "description": "Date range used (ISO dates)."
          }
        }
      },
...
      "MonthlyReportItem": {
        "type": "object",
        "description": "A single month's income/expense totals for the monthly report.",
        "required": [
          "year",
          "month",
          "income_total",
          "expense_total",
          "balance"
        ],
        "properties": {
          "balance": {
            "type": "string",
            "description": "Balance = income − expense."
          },
          "expense_total": {
            "type": "string",
            "description": "Total expenses for the month."
          },
          "income_total": {
            "type": "string",
            "description": "Total income for the month."
          },
          "month": {
            "type": "integer",
            "format": "int32",
            "description": "Month (1-12)."
          },
          "year": {
            "type": "integer",
            "format": "int32",
            "description": "Year."
          }
        }
      },
      "MonthlyReportResponse": {
        "type": "object",
        "description": "Response for the monthly report.",
        "required": [
          "months"
        ],
        "properties": {
          "months": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonthlyReportItem"
            },
            "description": "One entry per month in the requested range (chronological order)."
          }
        }
      },
...
      "TrendPoint": {
        "type": "object",
        "description": "A single point in the trends report.",
        "required": [
          "month_label",
          "year",
          "month",
          "income_total",
          "expense_total",
          "net"
        ],
        "properties": {
          "expense_total": {
            "type": "string",
            "description": "Expenses for the month."
          },
          "income_total": {
            "type": "string",
            "description": "Income for the month."
          },
          "month": {
            "type": "integer",
            "format": "int32",
            "description": "Month (1-12)."
          },
          "month_label": {
            "type": "string",
            "description": "Human-friendly label (e.g., \"2026-04\")."
          },
          "net": {
            "type": "string",
            "description": "Net (income − expense) for the month."
          },
          "year": {
            "type": "integer",
            "format": "int32",
            "description": "Year."
          }
        }
      },
      "TrendsResponse": {
        "type": "object",
        "description": "Response for the trends report.",
        "required": [
          "trends"
        ],
        "properties": {
          "trends": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrendPoint"
            },
            "description": "Monthly points, chronological order."
          }
        }
      },
...
    }
  },
  "security": [
    {}
  ],
...
}